requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
absl-py=0.13.0=pypi_0
aiohttp=3.7.4.post0=pypi_0
async-timeout=3.0.1=pypi_0
attrs=21.2.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.10.0=pypi_0
ca-certificates=2021.5.30=ha878542_0
cachetools=4.2.2=pypi_0
certifi=2021.5.30=pypi_0
cffi=1.14.6=pypi_0
chardet=4.0.0=pypi_0
charset-normalizer=2.0.4=pypi_0
cloudpickle=1.6.0=pypi_0
cycler=0.10.0=pypi_0
cython=0.29.24=pypi_0
debugpy=1.4.1=pypi_0
decorator=5.0.9=pypi_0
dill=0.3.4=pypi_0
disentanglement-lib=1.4=pypi_0
easydict=1.9=pypi_0
entrypoints=0.3=pypi_0
filelock=3.3.2=pypi_0
fsspec=2021.8.1=pypi_0
future=0.18.2=pypi_0
gdown=4.2.0=pypi_0
gin-config=0.4.0=pypi_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.39.0=pypi_0
h5py=3.4.0=pypi_0
idna=3.2=pypi_0
imageio=2.9.0=pypi_0
ipdb=0.13.9=pypi_0
ipykernel=6.3.1=pypi_0
ipython=7.27.0=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.18.0=pypi_0
joblib=1.0.1=pypi_0
jupyter-client=7.0.2=pypi_0
jupyter-core=4.7.1=pypi_0
kiwisolver=1.3.2=pypi_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
libffi=3.3=h58526e2_2
libgcc-ng=11.1.0=hc902ee8_8
libgomp=11.1.0=hc902ee8_8
libstdcxx-ng=11.1.0=h56837e0_8
markdown=3.3.4=pypi_0
matplotlib=3.4.3=pypi_0
matplotlib-inline=0.1.2=pypi_0
multidict=5.1.0=pypi_0
ncurses=6.2=h58526e2_4
nest-asyncio=1.5.1=pypi_0
numpy=1.21.2=pypi_0
oauthlib=3.1.1=pypi_0
opencv-python=4.5.3.56=pypi_0
openssl=1.1.1k=h7f98852_1
packaging=21.0=pypi_0
pandas=1.3.2=pypi_0
parso=0.8.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.3.1=pypi_0
pip=21.2.4=pyhd8ed1ab_0
prompt-toolkit=3.0.20=pypi_0
protobuf=3.17.3=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.2=pypi_0
pycparser=2.20=pypi_0
pygments=2.10.0=pypi_0
pymunk=6.2.0=pypi_0
pyparsing=2.4.7=pypi_0
pysocks=1.7.1=pypi_0
python=3.8.10=h49503c6_1_cpython
python-dateutil=2.8.2=pypi_0
python_abi=3.8=2_cp38
pytorch-lightning=1.2.7=pypi_0
pytz=2021.1=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=22.2.1=pypi_0
readline=8.1=h46c0cb4_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.7.1=pypi_0
seaborn=0.11.2=pypi_0
setuptools=57.4.0=py38h578d9bd_0
simplejson=3.17.5=pypi_0
six=1.16.0=pypi_0
soupsieve=2.3=pypi_0
sqlite=3.36.0=h9cd32fc_0
tensorboard=2.6.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow-hub=0.12.0=pypi_0
tensorflow-probability=0.7.0=pypi_0
threadpoolctl=2.2.0=pypi_0
tk=8.6.11=h21135ba_0
toml=0.10.2=pypi_0
torch=1.11.0=pypi_0
torchmetrics=0.5.1=pypi_0
tornado=6.1=pypi_0
tqdm=4.62.2=pypi_0
traitlets=5.1.0=pypi_0
typing-extensions=3.10.0.2=pypi_0
urllib3=1.26.6=pypi_0
wcwidth=0.2.5=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.37.0=pyhd8ed1ab_1
xz=5.2.5=h516909a_1
yacs=0.1.8=pypi_0
yarl=1.6.3=pypi_0
zlib=1.2.11=h516909a_1010
